Store visual pose and trajectory depth as float32 millimetres (#2730)
Depth was documented as int16 millimetres almost everywhere and enforced
almost nowhere. That dtype matches the Gemini 335, which emits uint16
millimetres natively; it is wrong for the Gemini 305, whose depth is computed
from stereo disparity as float32 millimetres, so its fractional part was lost
to the int16 contract.

Nothing pinned the dtype on disk — both libraries write depth with
group[key] = arr, which infers the dtype from whatever the driver produced.
So what actually landed was uint16 on a 335, float32 on a 305, and int16
from a camera-missing fallback, while every docstring claimed int16.

This makes float32 millimetres the declared, uniform dtype for depth in both
the visual pose and visual trajectory libraries. Units stay millimetres — only
the dtype moves — so the 0-is-invalid sentinel and every existing threshold
keep their meaning.

+++ sdk/client.py
@@ -4027,7 +4027,7 @@ class Robot:
       resolved_depth = camera_response.depth
 
     if resolved_depth is None:
-      resolved_depth = np.zeros((*rgb_image.shape[:2], 1), dtype=np.int16)
+      resolved_depth = np.zeros((*rgb_image.shape[:2], 1), dtype=np.float32)
 
     pose = rpc_api.VisualPoseEntry(
         name=name,
